Adjustment.

+++ src/low/easy/relic_bn_mul_low.c
@@ -118,7 +118,7 @@ void bn_muln_low(dig_t *c, const dig_t *a, const dig_t *b, size_t size) {
 	}
 }
 
-void bn_muld_low(dig_t *c, const dig_t *a, int sa, const dig_t *b, int sb,
+void bn_muld_low(dig_t *c, const dig_t *a, size_t sa, const dig_t *b, size_t sb,
 		uint_t l, uint_t h) {
 	int i, j, ta;
 	const dig_t *tmpa, *tmpb;
